pyproject.toml
[project]
name = "train-data-bridge"
version = "0.1.0"
description = "Lightweight toolkit for processing mid-sized datasets on Vast.ai instances and storing curated shards on Cloudflare R2."
authors = [{name = "SJTU Data Bridge"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34.0",
    "datasets>=2.19.1",
    "requests>=2.31.0",
    "smart-open>=6.4.0",
    "typer>=0.9.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "litdata>=0.2.0",
    "webdataset>=0.2.86",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "pytest>=8.2.0",
]

[project.scripts]
train-data-bridge = "data_bridge.cli:app"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
